# Conflicts:
#	banco/sac_banco.sql
#	pom.xml
#	src/main/java/com/sac/backend/config/SecurityConfig.java
#	src/main/java/com/sac/backend/controllers/AdministradorController.java
#	src/main/java/com/sac/backend/controllers/DatasController.java
#	src/main/java/com/sac/backend/interfaces/AdministradorRepository.java
#	src/main/java/com/sac/backend/interfaces/Control.java
#	src/main/java/com/sac/backend/interfaces/DatasRepository.java
#	src/main/java/com/sac/backend/interfaces/ServiceInterface.java
#	src/main/java/com/sac/backend/models/DatasModel.java
#	src/main/java/com/sac/backend/security/JWTAuthenticationFilter.java
#	src/main/java/com/sac/backend/security/JWTAuthorizationFilter.java
#	src/main/java/com/sac/backend/security/JWTUtil.java
#	src/main/java/com/sac/backend/security/UserDetailsServiceImpl.java
#	src/main/java/com/sac/backend/services/AdministradorService.java
#	src/main/java/com/sac/backend/services/AgendamentoService.java
#	src/main/java/com/sac/backend/services/DatasService.java
#	src/main/resources/application.properties

+++ src/main/java/com/sac/backend/services/AgendamentoService.java
@@ -1,21 +1,20 @@
 package com.sac.backend.services;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import javax.mail.MessagingException;
-
 import com.sac.backend.helpers.EmailSender;
 import com.sac.backend.interfaces.AgendamentoRepository;
-import com.sac.backend.models.AgendamentoModel;
+import com.sac.backend.interfaces.ServiceInterface;
+import com.sac.backend.models.Agendamento;
+import com.sac.backend.models.Datas;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
-public class AgendamentoService {
+public class AgendamentoService implements ServiceInterface<Agendamento> {
 
     @Autowired
     private AgendamentoRepository agendamentoRepository;
@@ -23,28 +22,51 @@ public class AgendamentoService {
     @Autowired
     private EmailSender emailSender;
 
-    public List getAllAgendamentos() {
+    @Autowired
+    private DatasService datasService;
+   
+    @Override
+    public Agendamento create(Agendamento agendamento) {
+        emailSender.enviarEmailConfirmacao(agendamento);
+        emailSender.enviarEmailConfirmacaoAdmin(agendamento);
 
-        List agendamentos = new ArrayList<>();
+        // Pega a data usada no agendamento, atualiza seu status para 1 (utilizada) e atualiza no banco
+        Datas dataAgendamento = agendamento.getData();
+        dataAgendamento.setStatus(1);
+        datasService.update(dataAgendamento);
+
+        return agendamentoRepository.save(agendamento);
+    }
+
+    @Override
+    public Optional<Agendamento> findById(Long id) {
+        Optional<Agendamento> agendamento = agendamentoRepository.findById(id);
+        return agendamento;
+    }
+
+    @Override
+    public List<Agendamento> findAll() {
+        List<Agendamento> agendamentos = new ArrayList<>();
         agendamentoRepository.findAll().forEach(agendamentos::add);
 
         return agendamentos;
     }
 
-    public Optional<AgendamentoModel> getAgendamento(Long id) {
-        return agendamentoRepository.findById(id);
+    @Override
+    public boolean update(Agendamento obj) {
+        if (agendamentoRepository.existsById(obj.getId())) {
+            agendamentoRepository.save(obj);
+            return true;
+        }
+        return false;
     }
 
-    public void addAgendamento(AgendamentoModel agendamento) {
-        emailSender.enviarEmailConfirmacao();
-        agendamentoRepository.save(agendamento);
-    }
-
-    public void updateAgendamento(String id, AgendamentoModel agendamento) {
-        agendamentoRepository.save(agendamento);
-    }
-
-    public void deleteAgendamento(Long id) {
-        agendamentoRepository.deleteById(id);
+    @Override
+    public boolean delete(Long id) {
+        if (agendamentoRepository.existsById(id)) {
+            agendamentoRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
